add tokenId type gas optimization

+++ script/DeployAirdropApp.s.sol
@@ -12,7 +12,7 @@ contract DeployAirdropApp is Script {
 
         address delegationIndexerAddress = 0x03FfF50ef7C829B1aF69dcaAd257bD9426897f41; // Deployed in Mumbai
         address mockERC721Address = 0x9Fd12DfDe6eF24fa793C699941363d8E61E4c983; // Deployed in Mumbai
-        uint256 tokenId = 0;
+        uint32 tokenId = 0;
         uint256 airdropTokenAmount = 100e18;
         AirdropApp airdropApp = new AirdropApp(delegationIndexerAddress, mockERC721Address, tokenId, airdropTokenAmount);
         vm.stopBroadcast();
